Server: Fix falldamage overflow, initialize vectors and more
Main thing coming in this commit is fixed falldamage.
I never thought the damage dealt to the player could be bigger then
100 but it can. Meaning it will overflow HP if its uint8

+++ Source/Packets.h
@@ -18,7 +18,7 @@ void SendBlockAction(Server* server, uint8 playerID, uint8 actionType, int X, in
 void SendStateData(Server* server, uint8 playerID);
 void SendInputData(Server* server, uint8 playerID);
 void sendKillPacket(Server* server, uint8 killerID, uint8 playerID, uint8 killReason, uint8 respawnTime);
-void sendHP(Server* server, uint8 hitPlayerID, uint8 playerID, uint8 HPChange, uint8 type, uint8 killReason, uint8 respawnTime);
+void sendHP(Server* server, uint8 hitPlayerID, uint8 playerID, long HPChange, uint8 type, uint8 killReason, uint8 respawnTime);
 void SendPlayerState(Server* server, uint8 playerID, uint8 otherID);
 void SendMapStart(Server* server, uint8 playerID);
 void SendMapChunks(Server* server, uint8 playerID);
